edusaas-ui/backend/requirements.txt
Django==4.2.0
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
Pillow==10.0.0
python-decouple==3.8
gunicorn==21.2.0
whitenoise==6.5.0
psycopg2-binary==2.9.9
